+++ cmp/form/field/BaseFieldModel.ts
@@ -5,12 +5,19 @@
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
 import {HoistModel, managed, TaskObserver} from '@xh/hoist/core';
-import {genDisplayName, required, Rule, RuleLike, ValidationState} from '@xh/hoist/data';
+import {
+    genDisplayName,
+    required,
+    Rule,
+    RuleLike,
+    ValidationIssue,
+    ValidationState
+} from '@xh/hoist/data';
 import {action, bindable, computed, makeObservable, observable, runInAction} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
 import {executeIfFunction, withDefault} from '@xh/hoist/utils/js';
 import {createObservableRef} from '@xh/hoist/utils/react';
-import {compact, flatten, isEmpty, isEqual, isFunction, isNil} from 'lodash';
+import {compact, flatten, isEmpty, isEqual, isFunction, isNil, isString} from 'lodash';
 import {FormModel} from '../FormModel';
 
 export interface BaseFieldConfig {
@@ -95,7 +102,7 @@ export abstract class BaseFieldModel extends HoistModel {
     // containing any validation errors for the rule.  If validation for the rule has not
     // completed will contain null
     @observable
-    private _errors: string[][];
+    private validationIssues: ValidationIssue[][];
 
     @managed
     private validationTask = TaskObserver.trackLast();
@@ -118,7 +125,7 @@ export abstract class BaseFieldModel extends HoistModel {
         this._disabled = disabled;
         this._readonly = readonly;
         this.rules = this.processRuleSpecs(rules);
-        this._errors = this.rules.map(() => null);
+        this.validationIssues = this.rules.map(() => null);
     }
 
     //-----------------------------
@@ -175,12 +182,29 @@ export abstract class BaseFieldModel extends HoistModel {
     /** All validation errors for this field. */
     @computed
     get errors(): string[] {
-        return compact(flatten(this._errors));
+        return compact(
+            flatten(this.validationIssues).map(it => (it?.severity === 'error' ? it.message : null))
+        );
+    }
+
+    /** All validation warnings for this field. */
+    @computed
+    get warnings(): string[] {
+        return compact(
+            flatten(this.validationIssues).map(it =>
+                it?.severity === 'warning' ? it.message : null
+            )
+        );
     }
 
     /** All validation errors for this field and its sub-forms. */
     get allErrors(): string[] {
         return this.errors;
+    }
+
+    /** All validation warnings for this field and its sub-forms. */
+    get allWarnings(): string[] {
+        return this.warnings;
     }
 
     /**
@@ -202,7 +226,7 @@ export abstract class BaseFieldModel extends HoistModel {
 
         // Force an immediate 'Unknown' state -- the async recompute leaves the old state in place until it completed.
         // (We want that for a value change, but not reset/init)  Force the recompute only if needed.
-        this._errors.fill(null);
+        this.validationIssues.fill(null);
         wait().then(() => {
             if (!this.isValidationPending && this.validationState === 'Unknown') {
                 this.computeValidationAsync();
@@ -272,9 +296,14 @@ export abstract class BaseFieldModel extends HoistModel {
         return this.deriveValidationState();
     }
 
-    /** True if this field is confirmed to be Valid. */
+    /** True if this field is confirmed to be Valid (with or without warnings). */
     get isValid(): boolean {
-        return this.validationState === 'Valid';
+        return this.validationState === 'Valid' || this.validationState === 'ValidWithWarnings';
+    }
+
+    /** True if this field is confirmed to be Valid but has warnings. */
+    get isValidWithWarnings(): boolean {
+        return this.validationState === 'ValidWithWarnings';
     }
 
     /** True if this field is confirmed to be NotValid. */
@@ -339,13 +368,13 @@ export abstract class BaseFieldModel extends HoistModel {
         const promises = this.rules.map(async (rule, idx) => {
             const result = await this.evaluateRuleAsync(rule);
             if (runId === this.validationRunId) {
-                runInAction(() => (this._errors[idx] = result));
+                runInAction(() => (this.validationIssues[idx] = result));
             }
         });
         await Promise.all(promises);
     }
 
-    private async evaluateRuleAsync(rule): Promise<string[]> {
+    private async evaluateRuleAsync(rule: Rule): Promise<ValidationIssue[]> {
         if (this.ruleIsActive(rule)) {
             const promises = rule.check.map(async constraint => {
                 const {value, name, displayName} = this,
@@ -355,7 +384,9 @@ export abstract class BaseFieldModel extends HoistModel {
             });
 
             const ret = await Promise.all(promises);
-            return compact(flatten(ret));
+            return compact(flatten(ret)).map(issue =>
+                isString(issue) ? {message: issue, severity: 'error'} : issue
+            );
         }
         return [];
     }
@@ -367,10 +398,11 @@ export abstract class BaseFieldModel extends HoistModel {
     }
 
     protected deriveValidationState(): ValidationState {
-        const {_errors} = this;
+        const {errors, warnings, validationIssues} = this;
 
-        if (_errors.some(e => !isEmpty(e))) return 'NotValid';
-        if (_errors.some(e => isNil(e))) return 'Unknown';
+        if (!isEmpty(errors)) return 'NotValid';
+        if (validationIssues.some(e => isNil(e))) return 'Unknown';
+        if (!isEmpty(warnings)) return 'ValidWithWarnings';
         return 'Valid';
     }
 }