+++ data/impl/RecordValidator.ts
@@ -4,9 +4,16 @@
  *
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
-import {Field, Rule, StoreRecord, StoreRecordId, ValidationState} from '@xh/hoist/data';
+import {
+    Field,
+    Rule,
+    StoreRecord,
+    StoreRecordId,
+    ValidationIssue,
+    ValidationState
+} from '@xh/hoist/data';
 import {computed, observable, makeObservable, runInAction} from '@xh/hoist/mobx';
-import {compact, flatten, isEmpty, mapValues, values} from 'lodash';
+import {compact, flatten, isEmpty, isString, mapValues, values} from 'lodash';
 import {TaskObserver} from '../../core';
 
 /**
@@ -16,18 +23,24 @@ import {TaskObserver} from '../../core';
 export class RecordValidator {
     record: StoreRecord;
 
-    @observable.ref _fieldErrors: RecordErrorMap = null;
-    _validationTask = TaskObserver.trackLast();
-    _validationRunId = 0;
+    @observable.ref private fieldValidationIssues: RecordValidationIssueMap = null;
+    private validationTask = TaskObserver.trackLast();
+    private validationRunId = 0;
 
     get id(): StoreRecordId {
         return this.record.id;
     }
 
-    /** True if the record is confirmed to be Valid. */
+    /** True if the record is confirmed to be Valid (with or without warnings). */
     @computed
     get isValid(): boolean {
-        return this.validationState === 'Valid';
+        return this.validationState === 'Valid' || this.validationState === 'ValidWithWarnings';
+    }
+
+    /** True if the record is confirmed to be Valid but has warnings. */
+    @computed
+    get isValidWithWarnings(): boolean {
+        return this.validationState === 'ValidWithWarnings';
     }
 
     /** True if the record is confirmed to be NotValid. */
@@ -44,20 +57,36 @@ export class RecordValidator {
 
     /** Map of field names to field-level errors. */
     @computed.struct
-    get errors(): RecordErrorMap {
-        return this._fieldErrors ?? {};
+    get errors(): RecordValidationMessagesMap {
+        return mapValues(this.fieldValidationIssues ?? {}, issues =>
+            compact(issues.map(it => (it?.severity === 'error' ? it?.message : null)))
+        );
+    }
+
+    /** Map of field names to field-level warnings. */
+    @computed.struct
+    get warnings(): RecordValidationMessagesMap {
+        return mapValues(this.fieldValidationIssues ?? {}, issues =>
+            compact(issues.map(it => (it?.severity === 'warning' ? it?.message : null)))
+        );
     }
 
     /** Count of all validation errors for the record. */
     @computed
     get errorCount(): number {
-        return flatten(values(this._fieldErrors)).length;
+        return flatten(values(this.errors)).length;
+    }
+
+    /** Count of all validation warnings for the record. */
+    @computed
+    get warningCount(): number {
+        return flatten(values(this.warnings)).length;
     }
 
     /** True if any fields are currently recomputing their validation state. */
     @computed
     get isPending(): boolean {
-        return this._validationTask.isPending;
+        return this.validationTask.isPending;
     }
 
     private _validators = [];
@@ -71,7 +100,7 @@ export class RecordValidator {
      * Recompute validations for the record and return true if valid.
      */
     async validateAsync(): Promise<boolean> {
-        let runId = ++this._validationRunId,
+        let runId = ++this.validationRunId,
             fieldErrors = {},
             {record} = this,
             fieldsToValidate = record.store.fields.filter(it => !isEmpty(it.rules));
@@ -83,26 +112,29 @@ export class RecordValidator {
                 fieldErrors[field.name].push(result);
             });
         });
-        await Promise.all(promises).linkTo(this._validationTask);
+        await Promise.all(promises).linkTo(this.validationTask);
 
-        if (runId !== this._validationRunId) return;
+        if (runId !== this.validationRunId) return;
         fieldErrors = mapValues(fieldErrors, it => compact(flatten(it)));
 
-        runInAction(() => (this._fieldErrors = fieldErrors));
+        runInAction(() => (this.fieldValidationIssues = fieldErrors));
 
         return this.isValid;
     }
 
     /** The current validation state for the record. */
     getValidationState(): ValidationState {
-        const {_fieldErrors} = this;
-
-        if (_fieldErrors === null) return 'Unknown'; // Before executing any rules
-
-        return values(_fieldErrors).some(errors => !isEmpty(errors)) ? 'NotValid' : 'Valid';
+        if (this.fieldValidationIssues === null) return 'Unknown'; // Before executing any rules
+        if (this.errorCount) return 'NotValid';
+        if (this.warningCount) return 'ValidWithWarnings';
+        return 'Valid';
     }
 
-    async evaluateRuleAsync(record: StoreRecord, field: Field, rule: Rule): Promise<string[]> {
+    async evaluateRuleAsync(
+        record: StoreRecord,
+        field: Field,
+        rule: Rule
+    ): Promise<ValidationIssue[]> {
         const values = record.getValues(),
             {name, displayName} = field,
             value = record.get(name);
@@ -114,7 +146,9 @@ export class RecordValidator {
             });
 
             const ret = await Promise.all(promises);
-            return compact(flatten(ret));
+            return compact(flatten(ret)).map(issue =>
+                isString(issue) ? {message: issue, severity: 'error'} : issue
+            );
         }
     }
 
@@ -124,5 +158,6 @@ export class RecordValidator {
     }
 }
 
-/** Map of Field names to Field-level error lists. */
-export type RecordErrorMap = Record<string, string[]>;
+/** Map of Field names to Field-level ValidationIssue lists. */
+export type RecordValidationIssueMap = Record<string, ValidationIssue[]>;
+export type RecordValidationMessagesMap = Record<string, string[]>;