+++ CHANGELOG.md
@@ -31,6 +31,7 @@ this release, but is not strictly required.
   `react-grid-layout` v2+ (not common).
 * Modified `DashCanvasModel.containerPadding` to apply to the `react-grid-layout` div created by the
   library, instead of the Hoist-created containing div. This may affect printing layouts.
+* Enhanced `Field.rules` to support warnings, handled via `Grid` and `Form` validation APIs.
 
 ### 🎁 New Features
 
@@ -64,6 +65,10 @@ this release, but is not strictly required.
   and restored by the browser).
 * Introduced opt-in `Grid` performance optimizations on an experimental basis with
   `GridExperimentalFlags.deltaSort` and `GridExperimentalFlags.disableScrollOptimization`
+
+### ⚙️ Typescript API Adjustments
+
+* Removed `RecordErrorMap` type (not expected to impact most applications).
 
 ### 📚 Libraries